industry: groundwork for further Crius updates
These changes are not advisable without also updating the SDE, and
more investigation is needed to make sure the code will work right
with new handling of ME, wastage, damage per job, and extra materials
in the new SDE. This commit is not ready to be pushed live yet.

Also we need to make sure the Crius SDE won't break anything else.