industry: use Steve Ronuken's blueprint tables
Steve made new tables to hold the blueprints.yaml data that CCP moved
out of invBlueprintTypes. (Just using the YAML is not practical as the
file is enormous and PyYAML blows itself away trying to load it.) Other
people have recreated the legacy tables, see for example
https://forums.eveonline.com/default.aspx?g=posts&t=360102&p=2 instead.
For now Steve's conversion is the de facto standard so I'm sort of
inclined to convert our stuff to it if it's not too painful, but we'll
see.